// src/client/models/display-information.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Navitia API - SNCF
 * Complete API documentation for Navitia/SNCF public transportation API. Based on https://doc.navitia.io
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface DisplayInformation {
    /**
     * Direction
     */
    'direction'?: string;
    /**
     * Code
     */
    'code'?: string;
    /**
     * Network name
     */
    'network'?: string;
    /**
     * Trip short name
     */
    'trip_short_name'?: string;
    /**
     * Headsign (destination or train number)
     */
    'headsign'?: string;
    /**
     * Label
     */
    'label'?: string;
    /**
     * Color (hex)
     */
    'color'?: string;
    /**
     * Text color (hex)
     */
    'text_color'?: string;
    /**
     * Commercial mode name
     */
    'commercial_mode'?: string;
    /**
     * Physical mode name
     */
    'physical_mode'?: string;
    /**
     * Description
     */
    'description'?: string;
    /**
     * List of equipment types
     */
    'equipments'?: Array<string>;
    /**
     * Route ID
     */
    'route_id'?: string;
    /**
     * Line ID
     */
    'line_id'?: string;
    /**
     * Display name
     */
    'name'?: string;
}